requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
Werkzeug==3.0.4
